+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Security in case of vulnerabilities.
 
 ### Added
 
+- Added `Deed Of Termination For the Church Supplemental Agreement` task for transfer project.
+
 ### Changed
 
 ### Fixed