+++ CHANGELOG.md
@@ -26,6 +26,8 @@ Security in case of vulnerabilities.
 
 ### Fixed
 
+- Fixed `Significant date` validation issue of `Stakeholder kick-off` task.
+
 ### Security
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-10-27.1000...main) for everything awaiting release