+++ CHANGELOG.md
@@ -18,6 +18,21 @@ Security in case of vulnerabilities.
 
 ### Added
 
+### Changed
+
+### Fixed
+
+### Security
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-10-29.1020...main) for everything awaiting release
+
+---
+
+## [1.25.0](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-10-29.1020) - 2025-10-29
+
+### Added
+
+- Added `Deed of termination for the master funding agreement` task for transfer project.
 - Added `Deed Of Termination For the Church Supplemental Agreement` task for transfer project.
 - Added `Delete project` functionality.
 - Added `Confirm the school has completed all actions` task for conversion project.
@@ -31,7 +46,7 @@ Security in case of vulnerabilities.
 
 ### Security
 
-See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-10-27.1000...main) for everything awaiting release
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-10-27.1000...production-2025-10-29.1020) for everything in the release
 
 ---
 
@@ -44,7 +59,6 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - Added transfer creation end point for prepare to complete
 - Added `Confirm if the bank details for the general annual grant payment need to change` task for transfer project.
 - Added  'Confirm the academy name' task for conversion project.
-- Added `Deed of termination for the master funding agreement` task for transfer project.
 
 - Added 'Land consent letter task for transfer projects'
 - Adding key contact record on confirming project handover and logging error message if key contact is already been added.